+++ include/iot/data.h
@@ -1761,13 +1761,32 @@ extern bool iot_typecode_equal (const iot_typecode_t * tc1, const iot_typecode_t
 
 /**
  * @brief Converts a vector to an array, vector elements must be castable to the target array type, vector elements
- * that cannot be cast are ignored. If no vector elements can be cast to the required type NULL is returned.
+ * that cannot be cast are ignored. If no vector elements can be cast to the required type an empty array is returned.
  *
  * @param vector The vector to transform
  * @param type   The data element type for the array
- * @return       The newly created array containing the vector elements, or NULL if vector could not be converted
+ * @return       The newly created array containing the vector elements, may be empty
  */
 extern iot_data_t * iot_data_vector_to_array (const iot_data_t * vector, iot_data_type_t type);
+
+/**
+ * @brief Transforms one basic data type to another, if value is castable.
+ *
+ * @param data   The data to transform
+ * @param type   The type for the transformed data
+ * @return       The newly created data or NULL if type/value could not be cast
+ */
+iot_data_t * iot_data_transform (const iot_data_t * data, iot_data_type_t type);
+
+/**
+ * @brief Transforms an array of one type to an array of another. Array element values must be castable from one type/value
+ * to another, or the array element is ignored. An empty array is returned if no array elements can be transformed.
+ *
+ * @param vector The array to transform
+ * @param type   The data element type for the transformed array
+ * @return       The newly created array containing the transformed array elements, may be empty
+ */
+extern iot_data_t * iot_data_array_transform (const iot_data_t * array, iot_data_type_t type);
 
 /**
  * @brief Returns the size of the contained data type in bytes.